Move event payloads to a submodule
